Refactor parent names to allow more countries
